WIP - changed file names; night_reader will be an I/O management file like night_writer.
